Better model

+++ src/model/user.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Debug, Default)]
 pub struct User {
     pub username: String,
     pub vanity: String,
@@ -10,7 +10,7 @@ pub struct User {
     pub birthdate: Option<String>,
     pub verified: bool,
     pub deleted: bool,
-    pub flags: u32,
+    pub flags: u8,
     #[serde(skip_serializing)]
     pub(crate) password: Option<String>,
     #[serde(skip_serializing)]
